// web/src/api/generated/NodeTxDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Hash32 } from "./Hash32";
import type { NodeId } from "./NodeId";
import type { NodeTxKind } from "./NodeTxKind";
import type { Outpoint } from "./Outpoint";
import type { Tier } from "./Tier";

/**
 * Fluxnode fields of a start/confirm transaction.
 */
export type NodeTxDto = { txid: Hash32, kind: NodeTxKind, collateral: Outpoint, endpoint: string | null, benchmark_tier: Tier | null, sig_time_ms: number, tx_version: number, p2sh: boolean, node: NodeId | null, };