// web/src/api/generated/TxInputDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Amount } from "./Amount";
import type { Hash32 } from "./Hash32";

/**
 * A transaction input.
 */
export type TxInputDto = { coinbase: boolean, prev_txid: Hash32 | null, prev_vout: number | null, address: string | null, value: Amount | null, };